dia1/techadvisor/requirements.txt
langchain
langgraph
openai
python-dotenv
langchain-openai
langchain-community